Changing Order of Functions
I made the list of functions in README.md in the same order as they appear in php-riot-api-php.
Consider rearranging the functions to follow the same structure as they appear in the actual Riot API website.
Also, it would be much appreciated if the functions were better documented by comments.  I'm available to help you write documentation.